Factor out the type interface for shadow and null
Introduce a separate type interface for types that will be handled by
AD. This interface currently allows for defining the (unique) shadow
type of the given type and for building a null value of the type. This
is a step towards removing the dependency on the arithmetic dialect from
the AD pass.