Drop the dependency on FuncOps from GradientUtils
We don't need to know the specific op class of a function-like op to
clone it. This makes the autodiff interface independent of the func
dialect, but the AD pass still depends on it because it creates a
function call. There is currently no way of creating a function call in
an abstracted way, short of which the pass dependency on the func
dialect cannot be removed.